I fix the bot
