+++ cartoonfilter_ds_projec/TelegrammBot/settings.py
@@ -1,0 +1,8 @@
+BOT_TOKEN = "Token from BotFather"
+PROXY = {
+    'proxy_url': 'socks5://t1.learn.python.ru:1080',
+    'urllib3_proxy_kwargs': {
+        'username': 'learn', 
+        'password': 'python'
+    }
+}
